requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
docxtpl==0.11.5
python-docx==1.0.1
pydantic==1.10.7 
python-dotenv==1.0.0
pillow==9.5.0
httpx==0.28.1
dotenv==0.9.9
requests==2.32.3
aiohttp==3.9.3
aiofiles==23.2.1
asyncio==3.4.3
PyPDF2==3.0.1
pandas==2.0.3
openpyxl==3.1.2
tabula-py==2.9.0
pytesseract==0.3.10
pdf2image==1.16.3
pdfplumber==0.10.2
PyMuPDF>=1.23.0
opencv-python==4.11.0.86
numpy>=1.21.0
tqdm>=4.64.0
urllib3>=1.26.0
curl_cffi>=0.9.0
pdfminer.six==20221105
beautifulsoup4==4.12.2
